refactor(tasks-new): switch to explicit flow types
