+++ flow-typed/box-ui-elements.js
@@ -18,6 +18,7 @@ import type { ContentSidebarProps } from '../src/elements/content-sidebar';
 import type { ContentOpenWithProps } from '../src/elements/content-open-with';
 import type { ContentPreviewProps } from '../src/elements/content-preview';
 import type { FeatureConfig } from '../src/elements/common/feature-checking';
+import type { TaskNew } from '../src/elements/content-sidebar/activity-feed/task-new/flowTypes';
 import {
     ACCESS_OPEN,
     ACCESS_COLLAB,
@@ -632,127 +633,6 @@ type Tasks = {
     entries: Array<Task>,
     total_count: number,
 };
-
-/* New Task Types START */
-
-type ID = string;
-type ISODate = string;
-
-type UserMini = User;
-
-type FileMini = {
-    id: ID,
-    name: string,
-    type: 'file',
-};
-
-type FolderMini = {
-    id: ID,
-    name: string,
-    type: 'folder',
-};
-
-type TaskCollabStatus = 'APPROVED' | 'REJECTED' | 'COMPLETED' | 'NOT_STARTED';
-
-type TaskStatus = TaskCollabStatus | 'IN_PROGRESS';
-
-type TaskMini = {|
-    created_at: ISODate,
-    id: ID,
-    modified_at: ISODate,
-    status: TaskStatus,
-    type: 'task',
-|};
-
-type TaskCollabRole = 'CREATOR' | 'ASSIGNEE';
-
-type TaskCollab<R> = {|
-    completed_at?: ?ISODate,
-    id: ID,
-    role: R,
-    status: TaskCollabStatus,
-    target: UserMini,
-    task?: TaskMini,
-    type: 'task_collaborator',
-|};
-
-type TaskCollabCreator = TaskCollab<'CREATOR'>;
-
-type TaskCollabAssignee = {|
-    ...TaskCollab<'ASSIGNEE'>,
-    permissions: {|
-        can_delete: boolean,
-        can_update: boolean,
-    |},
-|};
-
-type TaskLink = {|
-    description?: string,
-    id: ID,
-    permissions: {|
-        can_delete: boolean,
-        can_update: boolean,
-    |},
-    target?: ?FileMini | ?FolderMini | ?UserMini,
-    task?: TaskMini,
-    type: 'task_link',
-|};
-
-type MarkerPaginatedCollection<T> = {
-    entries: T[],
-    limit: number,
-    next_marker: ?string,
-};
-
-type TaskAssigneeCollection = MarkerPaginatedCollection<TaskCollabAssignee>;
-
-type TaskLinkCollection = MarkerPaginatedCollection<TaskLink>;
-
-// See https://github.com/facebook/flow/issues/7574
-// This is currently *not* enforcing the constant types
-// type TaskType = typeof TASK_TYPE_GENERAL | typeof TASK_TYPE_APPROVAL;
-// type TaskEditMode = typeof TASK_EDIT_MODE_CREATE | typeof TASK_EDIT_MODE_EDIT;
-type TaskType = 'GENERAL' | 'APPROVAL';
-type TaskEditMode = 'CREATE' | 'EDIT';
-
-type TaskNew = {|
-    assigned_to: TaskAssigneeCollection,
-    completed_at?: ?ISODate,
-    completion_rule?: 'ANY_ASSIGNEE' | 'ALL_ASSIGNEES',
-    created_at: ISODate,
-    created_by: TaskCollabCreator,
-    description: string,
-    due_at?: ?ISODate,
-    id: ID,
-    modified_at?: ISODate,
-    permissions: {|
-        can_create_task_collaborator: boolean,
-        can_create_task_link: boolean,
-        can_delete: boolean,
-        can_update: boolean,
-    |},
-    progress_at?: ?ISODate,
-    status: TaskStatus,
-    task_links: TaskLinkCollection,
-    task_type: TaskType,
-    type: 'task',
-|};
-
-type TaskPayload = {
-    description: string,
-    due_at?: ?string,
-    task_type: TaskType,
-};
-
-type TaskUpdatePayload = {
-    addedAssignees: SelectorItems,
-    description: string,
-    due_at?: ?string,
-    id: string,
-    removedAssignees: Array<TaskCollabAssignee>,
-};
-
-/* New Task Types END */
 
 type Comment = {
     created_at: string,
