+++ src/api/Feed.js
@@ -32,6 +32,14 @@ import {
     TASK_NEW_NOT_STARTED,
     TYPED_ID_FEED_PREFIX,
 } from '../constants';
+import type {
+    TaskCollabAssignee,
+    TaskCollabStatus,
+    TaskLink,
+    TaskNew,
+    TaskType,
+} from '../elements/content-sidebar/activity-feed/task-new/flowTypes';
+import type { TaskPayload, TaskUpdatePayload } from './flowTypes';
 
 const TASK_NEW_INITIAL_STATUS = TASK_NEW_NOT_STARTED;
 const TASK = 'task';
