+++ src/common/constants.js
@@ -1,5 +1,6 @@
-const ITEM_TYPE_FILE = 'file';
-const ITEM_TYPE_FOLDER = 'folder';
-const ITEM_TYPE_WEBLINK = 'web_link';
+// @flow
+const ITEM_TYPE_FILE: 'file' = 'file';
+const ITEM_TYPE_FOLDER: 'folder' = 'folder';
+const ITEM_TYPE_WEBLINK: 'web_link' = 'web_link';
 
 export { ITEM_TYPE_FILE, ITEM_TYPE_FOLDER, ITEM_TYPE_WEBLINK };
