+++ src/elements/content-sidebar/ActivitySidebar.js
@@ -23,6 +23,8 @@ import AddTaskButton from './AddTaskButton';
 import SidebarUtils from './SidebarUtils';
 import { DEFAULT_COLLAB_DEBOUNCE, ORIGIN_ACTIVITY_SIDEBAR, SIDEBAR_VIEW_ACTIVITY } from '../../constants';
 import './ActivitySidebar.scss';
+import type { TaskType, TaskNew } from './activity-feed/task-new/flowTypes';
+import type { TaskUpdatePayload } from '../../api/flowTypes';
 
 type ExternalProps = {
     currentUser?: User,
