+++ src/elements/content-sidebar/TaskModal.js
@@ -8,6 +8,7 @@ import TaskForm from './activity-feed/task-form';
 import messages from '../common/messages';
 import { TASK_EDIT_MODE_CREATE, TASK_TYPE_APPROVAL, TASK_TYPE_GENERAL } from '../../constants';
 import type { TaskFormProps } from './activity-feed/task-form/TaskForm';
+import type { TaskType, TaskEditMode } from './activity-feed/task-new/flowTypes';
 
 type TaskModalProps = {
     editMode?: TaskEditMode,
