+++ src/elements/content-sidebar/activity-feed/task-form/TaskForm.js
@@ -17,11 +17,13 @@ import Button from '../../../../components/button/Button';
 import PrimaryButton from '../../../../components/primary-button/PrimaryButton';
 import InlineError from '../../../../components/inline-error/InlineError';
 import { TASK_EDIT_MODE_CREATE, TASK_EDIT_MODE_EDIT } from '../../../../constants';
-
 import messages from '../../../common/messages';
 import { ACTIVITY_TARGETS, INTERACTION_TARGET } from '../../../common/interactionTargets';
 
 import './TaskForm.scss';
+
+import type { TaskCollabAssignee, TaskType, TaskEditMode } from '../task-new/flowTypes';
+import type { TaskUpdatePayload } from '../../../../api/flowTypes';
 
 type TaskFormProps = {|
     error?: any,
