+++ src/elements/content-sidebar/activity-feed/task-new/AvatarGroup.js
@@ -13,6 +13,9 @@ import { TASK_NEW_APPROVED, TASK_NEW_REJECTED, TASK_NEW_COMPLETED, TASK_NEW_NOT_
 
 import './AvatarGroup.scss';
 
+import type { TaskAssigneeCollection } from './flowTypes';
+import type { ISODate } from '../../../../common/flowTypes';
+
 const MAX_AVATARS = 3;
 
 type Props = {|
