+++ src/elements/content-sidebar/activity-feed/task-new/TaskDueDate.js
@@ -10,6 +10,9 @@ import { TASK_NEW_NOT_STARTED } from '../../../../constants';
 
 import './TaskDueDate.scss';
 
+import type { TaskStatus } from './flowTypes';
+import type { ISODate } from '../../../../common/flowTypes';
+
 type Props = {
     dueDate: ISODate,
     status: TaskStatus,
