+++ src/features/unified-share-modal/SharedLinkAccessLabel.js
@@ -2,7 +2,7 @@
 import * as React from 'react';
 import { FormattedMessage } from 'react-intl';
 
-import type { itemType as ItemType } from '../../common/box-types';
+import type { ItemType } from '../../common/flowTypes';
 
 import { ANYONE_WITH_LINK, ANYONE_IN_COMPANY, PEOPLE_IN_ITEM } from './constants';
