+++ src/features/unified-share-modal/SharedLinkAccessMenu.js
@@ -7,7 +7,7 @@ import DropdownMenu, { MenuToggle } from '../../components/dropdown-menu';
 import { Menu, SelectMenuItem } from '../../components/menu';
 import PlainButton from '../../components/plain-button';
 import Tooltip from '../../components/tooltip';
-import type { itemType as ItemType } from '../../common/box-types';
+import type { ItemType } from '../../common/flowTypes';
 
 import SharedLinkAccessLabel from './SharedLinkAccessLabel';
 import { ANYONE_WITH_LINK, ANYONE_IN_COMPANY, PEOPLE_IN_ITEM } from './constants';
